desk_arcive/requirements.txt
# Core Dependencies - 2025.12.20 최신화
# 자동 업그레이드: pip install --upgrade -r requirements.txt

requests>=2.32.5
beautifulsoup4>=4.14.3
feedparser
newspaper3k
firebase-admin
python-dotenv
lxml_html_clean
flask>=3.1.2
aiohttp>=3.13.2
playwright>=1.57.0
google-genai>=1.56.0
